Moved manager.ServiceFactory -> impl.ServiceFactory BlackListLogic class have been refactored Removed redundant methods in HibernateSessionFactory Renamed HibernateSessionFactory::getSession -> openSession
