+++ src/main/java/com/epam/internetshop/controllers/logic/LoginLogic.java
@@ -2,7 +2,7 @@ package com.epam.internetshop.controllers.logic;
 
 import com.epam.internetshop.domain.User;
 import com.epam.internetshop.services.UserService;
-import com.epam.internetshop.services.manager.ServiceFactory;
+import com.epam.internetshop.services.impl.ServiceFactory;
 import com.epam.internetshop.services.validator.UserValidator;
 
 public class LoginLogic {
