+++ src/main/java/com/epam/internetshop/controllers/logic/RegisterLogic.java
@@ -2,7 +2,7 @@ package com.epam.internetshop.controllers.logic;
 
 import com.epam.internetshop.domain.User;
 import com.epam.internetshop.services.UserService;
-import com.epam.internetshop.services.manager.ServiceFactory;
+import com.epam.internetshop.services.impl.ServiceFactory;
 
 public class RegisterLogic {
     private static final UserService userService = ServiceFactory.newUserService();
