+++ src/main/java/com/epam/internetshop/services/impl/PaymentServiceImpl.java
@@ -5,14 +5,12 @@ import com.epam.internetshop.DAO.PaymentDAO;
 import com.epam.internetshop.DAO.ProductDAO;
 import com.epam.internetshop.DAO.UserDAO;
 import com.epam.internetshop.DAO.impl.HibernateDAOFactory;
-import com.epam.internetshop.DAO.impl.PaymentDAOImpl;
 import com.epam.internetshop.domain.Payment;
 import com.epam.internetshop.domain.Product;
 import com.epam.internetshop.domain.User;
 import com.epam.internetshop.services.PaymentService;
 import com.epam.internetshop.services.exception.ProductException;
 import com.epam.internetshop.services.exception.UserException;
-import com.epam.internetshop.services.manager.ServiceFactory;
 import com.epam.internetshop.services.validator.PaymentValidator;
 import com.epam.internetshop.services.validator.ProductValidator;
 import com.epam.internetshop.services.validator.UserValidator;
