+++ src/main/webapp/index.jsp
@@ -1,6 +1,6 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@ page language="java" contentType="text/html;
-charset=UTF-8" pageEncoding="UTF-8" import="com.epam.internetshop.services.manager.ServiceFactory" import="com.epam.internetshop.domain.Product"%>
+charset=UTF-8" pageEncoding="UTF-8" import="com.epam.internetshop.services.impl.ServiceFactory" import="com.epam.internetshop.domain.Product"%>
 <html>
 <head>
 <title>Electronic Store. EPAM Lab</title>
