+++ src/main/webapp/logined.jsp
@@ -1,6 +1,6 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@ page contentType="text/html;charset=UTF-8" pageEncoding="UTF-8"
-         import="com.epam.internetshop.services.manager.ServiceFactory"
+         import="com.epam.internetshop.services.impl.ServiceFactory"
          import="java.util.HashMap"
          import="com.epam.internetshop.domain.Product" %>
 <html lang="en">
